requirements.txt
Cython
Pillow
PyYAML>=5.3
scipy>=1.4.1
seaborn>=0.11.0
tqdm>=4.41.0

--find-links https://download.pytorch.org/whl/cu101/torch_stable.html
torch == 1.8.0
torchvision == 0.9.0